+++ tests/component/readers/test_analog_unscaled_reader.py
@@ -8,7 +8,7 @@ import pytest
 
 import nidaqmx
 from nidaqmx.stream_readers import AnalogUnscaledReader
-from ..conftest import (
+from .conftest import (
     VOLTAGE_CODE_EPSILON,
     _get_voltage_code_offset_for_chan,
 )
